Create intial folder stracture setup route and some component to start our project
