+++ src/App.js
@@ -1,11 +1,19 @@
-import logo from './logo.svg';
-import './App.css';
+import { BrowserRouter, Route, Routes } from "react-router-dom";
+import "./App.css";
+import About from "./Pages/Home/About/About";
+import Contact from "./Pages/Home/Contact/Contact";
+import Home from "./Pages/Home/Home/Home";
 
 function App() {
   return (
-    <div className="App">
-      
-    </div> 
+    <BrowserRouter>
+      <Routes>
+        <Route path="/" element={<Home />} />
+        <Route path="/home" element={<Home />} />
+        <Route path="/about" element={<About />} />
+        <Route path="/contact" element={<Contact />} />
+      </Routes>
+    </BrowserRouter>
   );
 }
