--- src/Pages/About/About.js
@@ -1,11 +1,0 @@
-import React from 'react';
-
-const About = () => {
-    return (
-        <div>
-            
-        </div>
-    );
-};
-
-export default About;
