--- src/Pages/Contact/Contact.js
@@ -1,11 +1,0 @@
-import React from 'react';
-
-const Contact = () => {
-    return (
-        <div>
-            
-        </div>
-    );
-};
-
-export default Contact;
