+++ src/Pages/Home/About/About.js
@@ -1,0 +1,11 @@
+import React from "react";
+
+const About = () => {
+  return (
+    <div>
+      <h1>Here is our team members</h1>
+    </div>
+  );
+};
+
+export default About;
