+++ src/Pages/Home/Contact/Contact.js
@@ -1,0 +1,11 @@
+import React from "react";
+
+const Contact = () => {
+  return (
+    <div>
+      <h1>Touch | your need</h1>
+    </div>
+  );
+};
+
+export default Contact;
