+++ src/Pages/Home/Home/Home.js
@@ -1,0 +1,11 @@
+import React from "react";
+
+const Home = () => {
+  return (
+    <div>
+      <h1>This is Home Page</h1>
+    </div>
+  );
+};
+
+export default Home;
